requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
DateTime==4.7
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
jsons==1.6.3
Pillow==9.2.0
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
typish==1.9.3
tzdata==2022.6
urllib3==1.26.12
zope.interface==5.5.0
